social_network/requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
colorama==0.4.6
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
